[arm64][clang] Silence a register size mismatch warning
Even though we only need 32 bits here, clang warns that we should be
using a "w" register in the inline assembly (which is not legal with
mrs/msr). Silence the warning by declaring the value as unsigned long.

(cherry picked from commit a0de5d88dfc67b3ba34c0455b1619e12e6cfccae)

+++ arch/arm64/include/arch/arch_ops.h
@@ -32,7 +32,7 @@ static inline void arch_disable_ints(void) {
 }
 
 static inline bool arch_ints_disabled(void) {
-    unsigned int state;
+    unsigned long state;
 
     __asm__ volatile("mrs %0, daif" : "=r"(state));
     state &= (1<<7);
@@ -51,7 +51,7 @@ static inline void arch_disable_fiqs(void) {
 }
 
 static inline bool arch_fiqs_disabled(void) {
-    unsigned int state;
+    unsigned long state;
 
     __asm__ volatile("mrs %0, daif" : "=r"(state));
     state &= (1<<6);
